+++ CHANGELOG.md
@@ -53,6 +53,7 @@ All notable changes to this project will be documented in this file.
 * Fix config file to properly eager load files ([#708](https://gitlab.com/yale-sdmp/vesta/issues/708)).
 * Fix 404 issue with font files ([#720](https://gitlab.com/yale-sdmp/vesta/issues/720)).
 * Fix redirect on DrawSuite update failure ([#710](https://gitlab.com/yale-sdmp/vesta/issues/710)).
+* Make lottery assignment and oversubscription draws securely random ([#698](https://gitlab.com/yale-sdmp/vesta/issues/698)).
 
 ### Added
 * Add suite unmerging ([#257](https://yale.githost.io/sdmp/rails/vesta/issues/257)).